Splitted heatmap series into several files.
