+++ ts/Series/Heatmap/HeatmapPointOptions.d.ts
@@ -1,0 +1,30 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type ScatterPointOptions from '../Scatter/ScatterPointOptions';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface HeatmapPointOptions extends ScatterPointOptions {
+    pointPadding?: number;
+    value?: (number|null);
+}
+
+export default HeatmapPointOptions;
