+++ ts/Series/TilemapSeries.ts
@@ -16,7 +16,9 @@
 import type Axis from '../Core/Axis/Axis';
 import type BBoxObject from '../Core/Renderer/BBoxObject';
 import type DataLabelOptions from '../Core/Series/DataLabelOptions';
+import type HeatmapPointOptions from './Heatmap/HeatmapPointOptions';
 import type HeatmapSeries from './Heatmap/HeatmapSeries';
+import type HeatmapSeriesOptions from './Heatmap/HeatmapSeriesOptions';
 import type Point from '../Core/Series/Point';
 import type { SeriesStatesOptions } from '../Core/Series/SeriesOptions';
 import type SVGPath from '../Core/Renderer/SVG/SVGPath';
